Migliorie generali

+++ src/main/java/com/example/warehouse/controller/OrderController.java
@@ -23,24 +23,33 @@ public class OrderController {
     private OrderRestService orderRestService;
 
     @GetMapping("/orders")
-    public List<Order> searchOrders(){
-        return orderRestService.searchOrders();
+    public List<Order> searchOrders(@RequestHeader(value="userToken") String token) {
+        if (orderRestService.checkToken(token, 2)) {
+            return orderRestService.searchOrders();
+        } else {
+            return new ArrayList<>();
+        }
     }
 
     @PostMapping("/orders/delete")
-    public ReturnWithMessage deleteOrder(@RequestBody String jsonString){
+    public ReturnWithMessage deleteOrder(@RequestBody String jsonString/*, @RequestHeader(value="userToken") String token*/) {
         Gson gson = new Gson();
         Map mapJson = gson.fromJson(jsonString, Map.class);
         Map<String, Object> map = mapJson;
         Double d = Double.parseDouble(map.get("id").toString());
         Long l = d.longValue();
-        if(map.containsKey("id"))
-            return orderRestService.deleteOrder(l);
-        return new ReturnWithMessage(false, "Id non ricevuto dal server");
+        String token = map.get("userToken").toString();
+        if (orderRestService.checkToken(token, 2)) {
+            if (map.containsKey("id"))
+                return orderRestService.deleteOrder(l);
+            return new ReturnWithMessage(false, "Id non ricevuto dal server");
+        }else{
+            return new ReturnWithMessage(false, "Non hai l'autorizzazione");
+        }
     }
 
     @PostMapping("/orders/confirm/{id}/{idPosition}")
-    public ReturnWithMessage confirmOrder(@PathVariable("id") long id, @PathVariable("idPosition") long idPosition, @RequestBody String jsonString){
+    public ReturnWithMessage confirmOrder(@PathVariable("id") long id, @PathVariable("idPosition") long idPosition, @RequestBody String jsonString) {
         Gson gson = new Gson();
         Map mapJson = gson.fromJson(jsonString, Map.class);
         Map<String, Object> map = mapJson;
@@ -48,7 +57,7 @@ public class OrderController {
     }
 
     @PostMapping("/orders/confirmNew/{id}")
-    public ReturnWithMessage confirmOrder(@PathVariable("id") long id, @RequestBody String jsonString){
+    public ReturnWithMessage confirmOrder(@PathVariable("id") long id, @RequestBody String jsonString) {
         Gson gson = new Gson();
         Map mapJson = gson.fromJson(jsonString, Map.class);
         Map<String, String> map = mapJson;
@@ -66,14 +75,13 @@ public class OrderController {
         Gson gson = new Gson();
         Map mapJson = gson.fromJson(jsonString, Map.class);
         Map<String, Object> map = mapJson;
-        /*Double d = Double.parseDouble(map.get("total").toString());
-        double total = d.doubleValue();*/
         Integer i = Integer.parseInt(map.get("quantity").toString());
         int quantity = i.intValue();
-        Long l = Long.parseLong(map.get("idProductSupplier").toString());
-        long idProductSupplier = l.longValue();
-        System.out.println(map.get("description").toString() +" " + idProductSupplier +" "+ quantity);
-        //return null;
-        return orderRestService.addOrder(map.get("description").toString(), idProductSupplier, quantity);
+        Long l = Long.parseLong(map.get("idSupplier").toString());
+        long idSupplier = l.longValue();
+        //l = Long.parseLong(map.get("idProduct").toString());
+        Double d = Double.parseDouble(map.get("idProduct").toString());
+        long idProduct = d.longValue();
+        return orderRestService.addOrder(map.get("description").toString(), idSupplier, idProduct, quantity);
     }
 }
